Tweaks splitted in files and new tweaks
