+++ build.sbt
@@ -29,6 +29,7 @@ libraryDependencies ++= Seq(
   aar(macroidRoot),
   aar(androidAppCompat),
   aar(androidCardView),
-  aar(androidRecyclerview))
+  aar(androidRecyclerview),
+  aar(androidDesign))
 
 licenses := Seq("Apache License, Version 2.0" -> url("http://www.apache.org/licenses/LICENSE-2.0.txt"))
