+++ project/Versions.scala
@@ -3,6 +3,6 @@ object Versions {
   val appV = "0.3-SNAPSHOT"
   val scalaV = "2.11.7"
   val androidPlatformV = "android-23"
-  val androidV = "23.1.1"
+  val androidV = "23.2.1"
   val macroidV = "2.0.0-M5-SNAPSHOT"
 }
