+++ src/main/scala/com/fortysevendeg/macroid/extras/FloatingActionButtonTweaks.scala
@@ -1,0 +1,39 @@
+/*
+ *
+ *   Copyright (C) 2015 47 Degrees, LLC http://47deg.com hello@47deg.com
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License"); you may
+ *   not use this file except in compliance with the License. You may obtain
+ *   a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package com.fortysevendeg.macroid.extras
+
+import android.content.res.ColorStateList
+import android.support.design.widget.FloatingActionButton
+import com.fortysevendeg.macroid.extras.ResourcesExtras._
+import macroid.{Tweak, ContextWrapper}
+
+
+object FloatingActionButtonTweaks {
+  type W = FloatingActionButton
+
+  def fbaColorResource(id: Int, rippleId: Int)(implicit contextWrapper: ContextWrapper) = Tweak[W] { view =>
+    view.setBackgroundTintList(contextWrapper.application.getResources.getColorStateList(id))
+    view.setRippleColor(resGetColor(rippleId))
+  }
+
+  def fbaColor(color: Int, rippleColor: Int)(implicit contextWrapper: ContextWrapper) = Tweak[W] { view =>
+    view.setBackgroundTintList(ColorStateList.valueOf(color))
+    view.setRippleColor(rippleColor)
+  }
+}
