+++ src/main/scala/com/fortysevendeg/macroid/extras/SeekBarTweaks.scala
@@ -1,0 +1,33 @@
+/*
+ *
+ *   Copyright (C) 2015 47 Degrees, LLC http://47deg.com hello@47deg.com
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License"); you may
+ *   not use this file except in compliance with the License. You may obtain
+ *   a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package com.fortysevendeg.macroid.extras
+
+import android.widget.SeekBar
+import android.widget.SeekBar.OnSeekBarChangeListener
+import macroid.Tweak
+
+object SeekBarTweaks {
+  type W = SeekBar
+
+  def sbMax(maxValue: Int): Tweak[W] = Tweak[W](_.setMax(maxValue))
+
+  def sbProgress(progressValue: Int): Tweak[W] = Tweak[W](_.setProgress(progressValue))
+
+  def sbOnSeekBarChangeListener(listener: OnSeekBarChangeListener): Tweak[W] = Tweak[W](_.setOnSeekBarChangeListener(listener))
+}
