+++ src/main/scala/com/fortysevendeg/macroid/extras/TableLayoutTweaks.scala
@@ -1,0 +1,36 @@
+/*
+ *
+ *   Copyright (C) 2015 47 Degrees, LLC http://47deg.com hello@47deg.com
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License"); you may
+ *   not use this file except in compliance with the License. You may obtain
+ *   a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package com.fortysevendeg.macroid.extras
+
+import android.view.View
+import android.widget.TableLayout
+import macroid.Tweak
+
+object TableLayoutTweaks {
+  type W = TableLayout
+
+  def tlLayoutMargins(value: Int): Tweak[View] = Tweak[View] {
+    view ⇒
+      val param = new TableLayout.LayoutParams(view.getLayoutParams)
+      param.setMargins(value, value, value, value)
+      view.setLayoutParams(param)
+  }
+
+  def tlStretchAllColumns(stretchAllColumns: Boolean): Tweak[W] = Tweak[W](_.setStretchAllColumns(stretchAllColumns))
+}
